backend/requirements.txt
pandas
numpy
flask
openai
python-dotenv
